requirements.txt
django==5.1.4
redis==5.2.1
python-dotenv==1.0.1
psycopg2-binary==2.9.10
djangorestframework==3.15.2
pillow==11.1.0
djangorestframework-simplejwt==5.3.1
django-cors-headers==4.6.0
drf-yasg==1.21.8
